Updated ind_resolve to handle ng_size_t dimensions.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11979 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/wrapper.c
@@ -7310,8 +7310,8 @@ void NclAddUserFuncs(void)
     args = NewArgs(2);
 
     dimsizes[0] = 1;
-    SetArgTemplate(args, nargs, "integer", 1, NclANY);  nargs++;
-    SetArgTemplate(args, nargs, "integer", 1, NclANY);  nargs++;
+    SetArgTemplate(args, nargs, "numeric", 1, NclANY);  nargs++;
+    SetArgTemplate(args, nargs, "numeric", 1, NclANY);  nargs++;
 
     NclRegisterFunc(ind_resolve_W, args, "ind_resolve", nargs);
